requirements.txt
asgiref==3.7.2
certifi==2023.7.22
distlib==0.3.7
Django==4.2.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.7
exceptiongroup==1.1.2
filelock==3.12.2
inflection==0.5.1
iniconfig==2.0.0
packaging==23.1
pipenv==2023.7.23
platformdirs==3.9.1
pluggy==1.2.0
psycopg2-binary==2.9.6
PyJWT==2.8.0
pytest==7.4.0
pytest-django==4.5.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0.1
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.7.1
uritemplate==4.1.1
virtualenv==20.24.2
virtualenv-clone==0.5.7
